split mish.py into kernel (mish.py) and test (test_mish.py)
- mish.py: keep kernel + adapter (mish_forward), remove test section
- test_mish.py: new file with golden_mish + main test block, imports mish_forward from mish.py
- verified: python test_mish.py all 4 configs PASS (fp16/fp32/bf16 + aligned/prime)
